+++ entity/player/player_entity.go
@@ -4,8 +4,8 @@ type PlayerEntity struct {
 	Id       string
 	Name     string
 	Photo    string
-	Height   int32
-	Weight   int32
+	Height   float32
+	Weight   float32
 	Age      int32
 	Position string
 	Number   int32
